I have also made some small corrections to Ana's blog on Code Smells Part II.
Please notice in particular that I have changed the names in one of the
code samples. Ana used the word 'retribution', which in English can only
mean 'punishment'; I have replaced it with 'remuneration', which I think
is the meaning she intended. I have done a search through the document to
ensure I have picked up all the uses of this word.

The other changes are very small.