Added option to serve spanish version
When trying to work on the become_an_apprentice page we were getting an error. `rake serve` only creates the english version. `rake es` creates a spanish version but with a limit of three posts. So the new added `rake servees` will create the spanish version with all the posts. And we need all the posts because the page shows three specific posts that happen not to be the three most recent posts