New craftsman page (#698)
* updating craftsmanpage and adding translation

* add button description to yaml file

* resolve merge conflicts

* Added apply form to individual pages

The apply forms that were present on the default careers form have been added to the individual crafstman and apprentice pages. Also, apprentice had the full page replaced with a copy of the crafstman one until we get the new version text

* New data on Apprentice page

We have reverted to a similar version as it was before. We have added data provided by David. There is an specifiction section on the en.yml file for it. Still needs to be added to the es.yml file

* add partial for 'life of apprentice' articles

* replace hard-coded blocks with included partials

* Fixed issue with background color on benefits

* Added interested button at the end.

Copied from the version at the top. I have changed the indentation, so is left aligned correctly on the document

* Changes on apprenticeship data

A few unused entries have been deleted. The structure has been copied to the es version, but still with the english language

* Added enrolment button at the bottom

* Replicated structure of benefits

We have replicated the same structure of benefits on the apprentice page as they are on the crafstman page. Because the benefits are nearly identical, we have just added the mentor benefit on the en.yml file and moved benefits outside of become_a_craftsman

* Added benefits to spanish yaml

For completion purposes, I have copied the same structure of benefits from the english yaml page into the spanish yaml page

* Added option to serve spanish version

When trying to work on the become_an_apprentice page we were getting an error. `rake serve` only creates the english version. `rake es` creates a spanish version but with a limit of three posts. So the new added `rake servees` will create the spanish version with all the posts. And we need all the posts because the page shows three specific posts that happen not to be the three most recent posts

* check that life_of_apprentice posts exist before including them

* update EN Apprenticeship program description

* update EN Craftsman benefits section

* fixing identation for the Apprentice page

* fix identation for the Become a Craftsman page

* fix margin bottom and top for the apprentice page

* fix structure of Life Of Apprentice blog posts

* Moved apply buttons

All buttons have been centered. The second button has been moved inside the div for the benefits (at the end)

* Changed the button for become_a_craftsman page

* add padding to Craftsman page benefits section

* increase margin between application button and role description

* Fixed regression of text

* Reformatted apply form

The form has been redone following the mockup and fields provided by David. Still some styling to be done and behaviour hasn't been changed yet

* Corrected indentation

* Fixed indentation

* Changed form action

Now the javascript that responds to the form should pick up correctly the new data from the form