Add mandatory fields to "Apply" form (#700)
* Moved and improved validation

The validation of mandatory fields has been moved to an earlier point. Also, now the error message will display which fields are incorrect

* Added mandatory fields text

* Added two sections as mandatory

I have added the Software Craftmanship and the About you sections as mandatory. Also, separated with a space the (*) from the text to make it aesthetically more pleasing

* Small refactoring to follow JS conventions

* Small refactor to put together similar functionality