Clean up build and run process.
After the confusion around how to run the website locally reported by the people that touched the website recently, including me, I’ve created a PR to simplify the run/build process.
If you know how to do your local thing (on-machine, without Docker) - it should still work. The new recommended approach though is just to use docker and I’ve cleaned up the README to reflect that. The README had a lot of outdated info that was not true/did not work anymore - deleted it.

The docker container start should be MUCH faster now as well, so should
be possible to have everything running pretty quickly.