Katas for Functional Calisthenics (#789)
* Initial post text

The basics of the post, including the reasons and the different katas

* Lots of modifications

Corrected all the links. Corrected titles of sections. Added cover
image (and attribution). Added all the rules for the bank
kata. Corrected the text on the 8 queens kata.

* Modifications for going live

Changed the date of the post and resized the image

* Renamed file to reflect the right date