Circle two (#1009)
* Temporal commit

* Intermediate work while trying to make it work

* New build system

Moved to circleci 2.0. The Dockerfile was modified so we could use our
own specific image but not with an experimental version of node.

All the changes to the ci were done following the circle ci 2.0
documentation and examples. Hopefully all works at the first attempt.

* removed 1.0 version

* Reorganization

Althought the build is still not working, I realized that probably
having the three steps like they were would mean separate
containers. Now we have the single container for prod and pre

* Separated reference and jobs

* Added reference to build

* Moved steps away from reference

Reference seems to be useful only for data,  not steps.

* Added aws cli

Forgot to have aws cli on the docker image

* Added aws to path

* Corrected casing