Made changes requested by Nicole
can we make amelia the author of the post
can we put the answer to the questions on the next line underneath the questions?
can we add a little more space between the intro paragraph and the first question?
can we update the last sentence? (link email and the work linkedin with her contact info)
For anyone that’s interested in learning more about Codurance please reach out to me via email or on LinkedIn.